Add auth_review

+++ review/views.py
@@ -1,13 +1,16 @@
 import json
 
+from django.views   import View
+from django.http    import HttpResponse, JsonResponse
+from datetime       import datetime,timedelta,timezone
+
 from .models        import Review, TravelObject, Age, ReviewTourProduct
 from product.models import TourProduct
-
-from django.views import View
-from django.http  import HttpResponse, JsonResponse
-from datetime     import datetime,timedelta,timezone
+from account.models import Account
+from account.utils  import login_requested
 
 class ReviewView(View):
+    @login_requested
     def post(self, request, product_id):
         try:
             if TourProduct.objects.filter(number=product_id).exists():
@@ -16,6 +19,7 @@ class ReviewView(View):
                 Review.objects.create(
                     content    = data['content'],
                     grade      = data['grade'],
+                    account    = request.agent
                 )
 
                 last_review = Review.objects.latest('created_at')
@@ -33,6 +37,7 @@ class ReviewView(View):
     def get(self, request, product_id):
         review_list = [{
             'id'      : review_data.review.id,
+            'name'    : review_data.review.account.username,
             'content' : review_data.review.content,
             'grade'   : review_data.review.grade,
             'date'    : review_data.review.created_at
@@ -41,27 +46,47 @@ class ReviewView(View):
         return JsonResponse({'Review_list' : review_list}, status=200)
 
 class ReviewDetail(View):
+    @login_requested
     def post(self, request, product_id, review_id):
         try:
             if TourProduct.objects.filter(number=product_id).exists() and Review.objects.filter(id = review_id).exists():
-                data      = json.loads(request.body)
-                edit_time = datetime.now(timezone.utc) 
+                if Review.objects.get(id = review_id).account == request.agent:
 
-                Review.objects.filter(id=review_id).update(content = data['content'], updated_at = edit_time)
+                    data      = json.loads(request.body)
+                    edit_time = datetime.now(timezone.utc) 
 
-                return HttpResponse(status=200)
+                    Review.objects.filter(id=review_id).update(
+                        content = data['content'], 
+                        updated_at = edit_time, 
+                    )
+
+                    return HttpResponse(status=200)
+
+                return JsonResponse({'message' : 'INVALID_USER'}, status=401)
+
             return JsonResponse({'message' : 'INVALID_PRODUCT_OR_REVIEW'}, status=200)
-        except KeyError:
-            return JsonResponse({'message' : 'INVALID_KEYS'}, status=400)           
 
-    def delete(self, request, product_id, review_id):
+        except KeyError:
+            return JsonResponse({'message' : 'INVALID_KEYS'}, status=400)
+
+    @login_requested
+    def delete(self, request, product_id, review_id): 
         try:
             if TourProduct.objects.filter(number=product_id).exists() and Review.objects.filter(id = review_id).exists():
-                deleted_review = ReviewTourProduct.objects.get(tour_product = TourProduct.objects.get(number = product_id).id, review = review_id)
-                deleted_review.delete()
+                if Review.objects.get(id = review_id).account == request.agent:
+                    deleted_review = ReviewTourProduct.objects.get(
+                        tour_product = TourProduct.objects.get(number = product_id).id, 
+                        review = review_id
+                    )
 
-                return HttpResponse(status=200)
+                    deleted_review.delete()
+
+                    return HttpResponse(status=200)
+                    
+                return JsonResponse({'message' : 'INVALID_USER'}, status=401)
+
             return JsonResponse({'message' : 'INVALID_PRODUCT_OR_REVIEW'}, status=200)
+
         except ReviewTourProduct.DoesNotExist:
             return HttpResponse(status=404)
